requirementx.txt
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
easygui==0.98.1
google-api-core==1.17.0
google-auth==1.14.1
google-cloud-core==1.3.0
google-cloud-speech==1.3.2
google-cloud-storage==1.28.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
idna==2.9
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydub==0.23.1
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
urllib3==1.25.9
